backend/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.2
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-magic>=0.4.27
python-dotenv>=1.0.0
alembic>=1.12.1
email-validator>=2.1.0

# Development dependencies
pytest>=7.4.3
pytest-cov>=4.1.0
httpx>=0.25.1
python-multipart>=0.0.6
python-magic-bin>=0.4.14; sys_platform == 'win32'
